CLN: move fred to test_data.py
